project_yolo/requirements.txt
# Core dependencies
ultralytics>=8.3.0
# torch и torchvision должны быть установлены отдельно с CUDA поддержкой
# НЕ устанавливайте их через pip install -r requirements.txt, если уже установлены с CUDA!
# Установка: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# torch>=2.0.0
# torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Visualization and metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Optional: GPU monitoring
pynvml>=11.5.0

# Optional: Experiment tracking
wandb>=0.15.0

# Optional: TensorRT (install separately if needed)
# tensorrt>=8.6.0
# pycuda>=2022.2.2

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

